LockContext was put in the wrong module. This is DC configuration class - it has to sit with the DC model. This is not a common TC structure.
